modules/ocrtesseract/requirements.txt
pillow
pytesseract
opencv-python